chore(types): organize request and response body types
Having all of the types in client.go felt wrong. Splitting
them into separate files for request and response helps us
keep track of the varying datatypes and minimize confusion.
It also made it easier/appropriate to add methods to those
types for data processing.

Semver: patch
