+++ logdna/client_test.go
@@ -12,10 +12,10 @@ import (
 
 func TestClient_View(t *testing.T) {
 	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		json.NewEncoder(w).Encode(ViewResponsePayload{ViewID: "test123456"})
+		json.NewEncoder(w).Encode(ViewResponse{ViewID: "test123456"})
 	}))
 	defer ts.Close()
-	payload := ViewPayload{Name: "test", Query: "test"}
+	payload := ViewRequest{Name: "test", Query: "test"}
 	client := Client{ServiceKey: servicekey, HTTPClient: ts.Client()}
 	result, err := client.CreateView(ts.URL, payload)
 	assert.Equal(t, nil, err)
@@ -27,9 +27,9 @@ func TestClient_Alert(t *testing.T) {
 		json.NewEncoder(w).Encode(AlertResponsePayload{PresetID: "test123456"})
 	}))
 	defer ts.Close()
-	var channels []Channel
-	channels = append(channels, Channel{Integration: "pagerduty", Key: "Your PagerDuty API key goes here", TriggerInterval: "15m", TriggerLimit: 20})
-	payload := ViewPayload{Name: "test", Channels: channels}
+	var channels []ChannelRequest
+	channels = append(channels, ChannelRequest{Integration: "pagerduty", Key: "Your PagerDuty API key goes here", TriggerInterval: "15m", TriggerLimit: 20})
+	payload := ViewRequest{Name: "test", Channels: channels}
 	client := Client{ServiceKey: servicekey, HTTPClient: ts.Client()}
 	result, err := client.CreateAlert(ts.URL, payload)
 	assert.Equal(t, nil, err)
@@ -39,7 +39,7 @@ func TestClient_Alert(t *testing.T) {
 func TestClient_ViewRequestError(t *testing.T) {
 	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {}))
 	defer ts.Close()
-	payload := ViewPayload{Name: "test", Query: "test"}
+	payload := ViewRequest{Name: "test", Query: "test"}
 	client := Client{ServiceKey: servicekey, HTTPClient: ts.Client()}
 	err := client.UpdateView(ts.URL, "\r\n", payload)
 	assert.Error(t, err)
@@ -51,7 +51,7 @@ func TestClient_ViewResponseError(t *testing.T) {
 		w.Write([]byte("{ this is invalid JSON }"))
 	}))
 	defer ts.Close()
-	payload := ViewPayload{Name: "test", Query: "test"}
+	payload := ViewRequest{Name: "test", Query: "test"}
 	client := Client{ServiceKey: servicekey, HTTPClient: ts.Client()}
 	_, err := client.CreateView(ts.URL, payload)
 	assert.Error(t, err)
@@ -60,7 +60,7 @@ func TestClient_ViewResponseError(t *testing.T) {
 func TestClient_AlertRequestError(t *testing.T) {
 	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {}))
 	defer ts.Close()
-	payload := ViewPayload{Name: "test"}
+	payload := ViewRequest{Name: "test"}
 	client := Client{ServiceKey: servicekey, HTTPClient: ts.Client()}
 	err := client.UpdateAlert(ts.URL, "\r\n", payload)
 	assert.Error(t, err)
@@ -72,7 +72,7 @@ func TestClient_AlertResponseError(t *testing.T) {
 		w.Write([]byte("{ this is invalid JSON }"))
 	}))
 	defer ts.Close()
-	payload := ViewPayload{Name: "test"}
+	payload := ViewRequest{Name: "test"}
 	client := Client{ServiceKey: servicekey, HTTPClient: ts.Client()}
 	_, err := client.CreateAlert(ts.URL, payload)
 	assert.Error(t, err)
@@ -87,7 +87,7 @@ func TestClient_ViewURLError(t *testing.T) {
 		json.NewEncoder(w).Encode(response)
 	}))
 	defer ts.Close()
-	payload := ViewPayload{Name: "test", Query: "test"}
+	payload := ViewRequest{Name: "test", Query: "test"}
 	client := Client{ServiceKey: servicekey, HTTPClient: ts.Client()}
 	_, err := client.CreateView(ts.URL, payload)
 	assert.Equal(t, errors.New("Invalid URL Error"), err)
@@ -102,7 +102,7 @@ func TestClient_AlertURLError(t *testing.T) {
 		json.NewEncoder(w).Encode(response)
 	}))
 	defer ts.Close()
-	payload := ViewPayload{Name: "test"}
+	payload := ViewRequest{Name: "test"}
 	client := Client{ServiceKey: servicekey, HTTPClient: ts.Client()}
 	_, err := client.CreateAlert(ts.URL, payload)
 	assert.Equal(t, errors.New("Invalid URL Error"), err)
@@ -117,7 +117,7 @@ func TestClient_ViewResourceNotFound(t *testing.T) {
 		json.NewEncoder(w).Encode(response)
 	}))
 	defer ts.Close()
-	payload := ViewPayload{Name: "test", Query: "test"}
+	payload := ViewRequest{Name: "test", Query: "test"}
 	client := Client{ServiceKey: servicekey, HTTPClient: ts.Client()}
 	_, err := client.CreateView(ts.URL, payload)
 	assert.Equal(t, errors.New("Resource Not Found"), err)
@@ -132,7 +132,7 @@ func TestClient_AlertResourceNotFound(t *testing.T) {
 		json.NewEncoder(w).Encode(response)
 	}))
 	defer ts.Close()
-	payload := ViewPayload{Name: "test"}
+	payload := ViewRequest{Name: "test"}
 	client := Client{ServiceKey: servicekey, HTTPClient: ts.Client()}
 	_, err := client.CreateAlert(ts.URL, payload)
 	assert.Equal(t, errors.New("Resource Not Found"), err)
