+++ logdna/resource_alert.go
@@ -21,7 +21,7 @@ func resourceAlertCreate(ctx context.Context, d *schema.ResourceData, m interfac
 	if err != nil {
 		return diag.FromErr(errors.New("bodytemplate json configuration is invalid"))
 	}
-	payload := ViewPayload{Name: name, Channels: channels}
+	payload := ViewRequest{Name: name, Channels: channels}
 	resp, err := client.CreateAlert(config.URL, payload)
 	if err != nil {
 		return diag.FromErr(err)
@@ -47,7 +47,7 @@ func resourceAlertUpdate(ctx context.Context, d *schema.ResourceData, m interfac
 	if err != nil {
 		return diag.FromErr(errors.New("bodytemplate json configuration is invalid"))
 	}
-	payload := ViewPayload{Name: name, Channels: channels}
+	payload := ViewRequest{Name: name, Channels: channels}
 	err = client.UpdateAlert(config.URL, presetID, payload)
 	if err != nil {
 		return diag.FromErr(err)
