+++ logdna/response_bodies.go
@@ -1,0 +1,38 @@
+package logdna
+
+// This separation of concerns between request and response bodies is only due
+// to inconsistencies in the API data types returned by the PUT versus the ones
+// returned by the GET. In a perfect world, they would use the same types.
+
+type ViewResponse struct {
+	Apps     []string          `json:"apps,omitempty"`
+	Category []string          `json:"category,omitempty"`
+	Channels []ChannelResponse `json:"channels,omitempty"`
+	Error    string            `json:"error,omitempty"`
+	Hosts    []string          `json:"hosts,omitempty"`
+	Levels   []string          `json:"levels,omitempty"`
+	Name     string            `json:"name,omitempty"`
+	Query    string            `json:"query,omitempty"`
+	Tags     []string          `json:"tags,omitempty"`
+	ViewID   string            `json:"viewID,omitempty"`
+}
+
+// ChannelResponse contains channel data returned from the logdna APIs
+// NOTE - Properties with `interface` are due to the APIs returning
+// some things as strings (PUT/emails) and other times arrays (GET/emails)
+type ChannelResponse struct {
+	AlertID         string            `json:"alertid,omitempty"`
+	BodyTemplate    string            `json:"bodyTemplate,omitempty"`
+	Emails          interface{}       `json:"emails,omitempty"`
+	Headers         map[string]string `json:"headers,omitempty"`
+	Immediate       bool              `json:"immediate"`
+	Integration     string            `json:"integration,omitempty"`
+	Key             string            `json:"key,omitempty"`
+	Method          string            `json:"method,omitempty"`
+	Operator        string            `json:"operator,omitempty"`
+	Terminal        bool              `json:"terminal,omitempty"`
+	TriggerInterval interface{}       `json:"triggerinterval,omitempty"`
+	TriggerLimit    int               `json:"triggerlimit,omitempty"`
+	Timezone        string            `json:"timezone,omitempty"`
+	URL             string            `json:"url,omitempty"`
+}
